change block height type from uint64 to uint32

+++ core/types/header.go
@@ -29,7 +29,7 @@ import (
 type Header struct {
 	Version          uint32
 	ShardID          uint64
-	ParentHeight     uint64
+	ParentHeight     uint32
 	PrevBlockHash    common.Uint256
 	TransactionsRoot common.Uint256
 	BlockRoot        common.Uint256
@@ -66,7 +66,7 @@ func (bd *Header) Serialization(sink *common.ZeroCopySink) error {
 func (bd *Header) serializationUnsigned(sink *common.ZeroCopySink) {
 	sink.WriteUint32(bd.Version)
 	sink.WriteUint64(bd.ShardID)
-	sink.WriteUint64(bd.ParentHeight)
+	sink.WriteUint32(bd.ParentHeight)
 	sink.WriteBytes(bd.PrevBlockHash[:])
 	sink.WriteBytes(bd.TransactionsRoot[:])
 	sink.WriteBytes(bd.BlockRoot[:])
@@ -143,7 +143,7 @@ func (bd *Header) deserializationUnsigned(source *common.ZeroCopySource) error {
 
 	bd.Version, eof = source.NextUint32()
 	bd.ShardID, eof = source.NextUint64()
-	bd.ParentHeight, eof = source.NextUint64()
+	bd.ParentHeight, eof = source.NextUint32()
 	bd.PrevBlockHash, eof = source.NextHash()
 	bd.TransactionsRoot, eof = source.NextHash()
 	bd.BlockRoot, eof = source.NextHash()
